\name{SA24204169-package}
\alias{SA24204169-package}
\docType{package}
\title{
  R package for DRO mean-var model and StatComp homework
}
\description{
 This package is for the DRO mean-var model and StatComp homework
}
\author{
Zhang chunxu

Maintainer: Zhang chunxu <1336033867@qq.com>
}
\keyword{ package }
