% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_w.R
\name{optimize_weights}
\alias{optimize_weights}
\title{A R-function for the DRO mean-variance optimization}
\usage{
optimize_weights(data, phi, n, delta, alpha)
}
\arguments{
\item{data}{Data matrix}

\item{phi}{Initial weight vector}

\item{n}{Number of data points to consider}

\item{delta}{Robustness parameter}

\item{alpha}{Risk-free rate}
}
\value{
List containing optimized returns and weight differences
}
\description{
This file contains the R functions for the DRO mean-variance optimization.
}
\examples{
\dontrun{
data <- matrix(rnorm(1000), ncol = 10)
phi <- rep(1/10, 10)
n <- 100
delta <- 0.1
alpha <- 0.01
optimize_weights(data, phi, n, delta, alpha)
}
}
