% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pvalue}
\alias{pvalue}
\title{A function to calculate the p-value for the Sharpe ratio}
\usage{
pvalue(r, r0)
}
\arguments{
\item{r}{Numeric vector of returns}

\item{r0}{Numeric vector of benchmark returns}
}
\value{
p-value for the Sharpe ratio
}
\description{
This function calculates the p-value for the Sharpe ratio
}
\examples{
\dontrun{
r <- rnorm(100)
r0 <- rnorm(100)
pvalue(r, r0)
}
}
